+++ app/gui/src/project-view/components/GraphEditor/GraphEdges.vue
@@ -14,6 +14,7 @@ import { type GraphNavigator } from '@/providers/graphNavigator'
 import { useGraphSelection } from '@/providers/graphSelection'
 import { injectInteractionHandler, type Interaction } from '@/providers/interactionHandler'
 import type { PortId } from '@/providers/portInfo'
+import { useOngoingAiPrompts } from '@/stores/ongoingAiPrompts'
 import { Ast } from '@/util/ast'
 import { isAstId, type AstId } from '@/util/ast/abstract'
 import { Vec2 } from '@/util/data/vec2'
@@ -32,6 +33,7 @@ const {
 } = useCurrentProject()
 const selection = useGraphSelection(true)
 const interaction = injectInteractionHandler()
+const aiPrompts = useOngoingAiPrompts()
 const connectionToast = useToast.error()
 
 const props = defineProps<{
@@ -242,7 +244,7 @@ function createNewNodeFromPort(id: AstId) {
         />
         <template v-for="id in nodeIdsWithOutputPorts" :key="id">
           <GraphNodeOutputPorts
-            v-show="id !== graph.editedNodeInfo?.id"
+            v-show="id !== graph.editedNodeInfo?.id && !aiPrompts.hiddenNodeIds.has(id)"
             :nodeId="id"
             @newNodeClick="(portId) => createNewNodeFromPort(portId)"
             @newNodeDrag="(portId) => graph.createEdgeFromNewButton(portId)"